cypherflux-backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
python-nmap==0.7.1
cryptography==41.0.7
psycopg2-binary==2.9.9
Werkzeug==3.0.1
sqlalchemy==2.0.23
